refactor: 컴포넌트화 시작
